Move globalVirtualEnvLocator.ts
